moved battleship test to dev agent and adjusted dev agent prompt
