+++ packages/agents/src/scriptedAgents/Developer/DeveloperAgent.spec.ts
@@ -17,6 +17,7 @@ import dotenv from "dotenv";
 import path from "path";
 import cl100k_base from "gpt-tokenizer/cjs/encoding/cl100k_base";
 import { ScriptedAgent, DeveloperAgent } from "..";
+import fs from "fs";
 
 const rootDir = path.join(__dirname, "../../../../../");
 
@@ -28,7 +29,7 @@ jest.setTimeout(120000);
 
 describe('Dev Agent Test Suite', () => {
 
-  function createDevAgent(testName: string): {
+  function createDevAgent(testName: string, pathsForFilesToInclude: string[] = []): {
     agent: ScriptedAgent;
     debugLog: DebugLog;
   } {
@@ -70,6 +71,15 @@ describe('Dev Agent Test Suite', () => {
     const scripts = new Scripts(scriptsWorkspace, "./");
 
     const workspace = new FileSystemWorkspace(testCaseDir);
+
+    for (const filePath of pathsForFilesToInclude) {
+      if (!fs.existsSync(filePath)) {
+        throw Error(`Input file does not exist: ${filePath}`);
+      }
+      const fileName = path.basename(filePath);
+      const fileContents = fs.readFileSync(filePath, "utf-8");
+      workspace.writeFileSync(fileName, fileContents);
+    }
 
     return {
       agent: new DeveloperAgent(
@@ -148,6 +158,30 @@ describe('Dev Agent Test Suite', () => {
 
     expect(response.value.ok).toBe(true);
     const sourceCode = agent.workspace.readFileSync("organize_files.py");
+    expect(sourceCode).toBeTruthy();
+  });
+
+  test("Battleship", async () => {
+    const { agent, debugLog } = createDevAgent(
+      "Battleship",
+      [
+        path.join(__dirname, "testInputs/Battleship/__init__.py"),
+        path.join(__dirname, "testInputs/Battleship/abstract_class.py"),
+        path.join(__dirname, "testInputs/Battleship/conftest.py"),
+        path.join(__dirname, "testInputs/Battleship/product_requirements.txt"),
+        path.join(__dirname, "testInputs/Battleship/test_negative.py"),
+        path.join(__dirname, "testInputs/Battleship/test_positive.py"),
+        path.join(__dirname, "testInputs/Battleship/user_stories.txt"),
+      ]
+    );
+    const response = await runDevAgent(
+      agent,
+      "Build a battleship game\n\nSpecifications:\n\nOverview: Battleship is a two-player strategy game where each player places their fleet of ships on a grid and tries to sink the opponent's fleet by guessing their locations.\nPlayers take turns calling out a row and column, attempting to name a square containing one of the opponent's ships.\n\nThe Grid: Each player's grid is a 10x10 grid, identified by rows (using numbers 1-10) and columns (using letters A-J).\n\nShips:\n\nCarrier - 5 squares\nBattleship - 4 squares\nCruiser - 3 squares\nSubmarine - 3 squares\nDestroyer - 2 squares\nEach ship occupies contiguous squares on the grid, arranged either horizontally or vertically.\n\nSetup:\n\nAt the start of the game, each player places their fleet on their grid. This setup is hidden from the opponent.\nThe game begins with Player 1, followed by Player 2, and so on.\nTaking Turns:\n\nOn a player's turn, they announce a grid square (e.g., \"D5\").\nThe opponent announces whether that square is a \"hit\" (if there's a part of a ship on that square) or \"miss\" (if the square is empty).\nIf a player hits a square occupied by a ship, they get another turn to guess. This continues until they make a miss, at which point their turn ends.\nIf a player hits all the squares occupied by a ship, the opponent must announce the sinking of that specific ship, e.g., \"You sank my Battleship!\"\n\nObjective: The goal is to sink all of your opponent's ships before they sink yours.\n\nEnd of the Game: The game ends when one player has sunk all of the opponent's ships. The winner is the player who sinks all the opposing fleet first.\n\nTechnical details:\nIn your root folder you will find an abstract class that defines the public interface of the Battleship class you will have to build:\n```\nfrom abc import ABC, abstractmethod\nfrom typing import Optional\n\nfrom pydantic import BaseModel, validator\n\n\n# Models for the request and response payloads\nclass ShipPlacement(BaseModel):\n    ship_type: str\n    start: dict  # {\"row\": int, \"column\": str}\n    direction: str\n\n    @validator(\"start\")\n    def validate_start(cls, start):\n        row, column = start.get(\"row\"), start.get(\"column\")\n\n        if not (1 <= row <= 10):\n            raise ValueError(\"Row must be between 1 and 10 inclusive.\")\n\n        if column not in list(\"ABCDEFGHIJ\"):\n            raise ValueError(\"Column must be one of A, B, C, D, E, F, G, H, I, J.\")\n\n        return start\n\n\nclass Turn(BaseModel):\n    target: dict  # {\"row\": int, \"column\": str}\n\n\nclass TurnResponse(BaseModel):\n    result: str\n    ship_type: Optional[str]  # This would be None if the result is a miss\n\n\nclass GameStatus(BaseModel):\n    is_game_over: bool\n    winner: Optional[str]\n\n\nfrom typing import List\n\n\nclass Game(BaseModel):\n    game_id: str\n    players: List[str]\n    board: dict  # This could represent the state of the game board, you might need to flesh this out further\n    ships: List[ShipPlacement]  # List of ship placements for this game\n    turns: List[Turn]  # List of turns that have been taken\n\n\nclass AbstractBattleship(ABC):\n    SHIP_LENGTHS = {\n        \"carrier\": 5,\n        \"battleship\": 4,\n        \"cruiser\": 3,\n        \"submarine\": 3,\n        \"destroyer\": 2,\n    }\n\n    @abstractmethod\n    def create_ship_placement(self, game_id: str, placement: ShipPlacement) -> None:\n        \"\"\"\n        Place a ship on the grid.\n        \"\"\"\n        pass\n\n    @abstractmethod\n    def create_turn(self, game_id: str, turn: Turn) -> TurnResponse:\n        \"\"\"\n        Players take turns to target a grid cell.\n        \"\"\"\n        pass\n\n    @abstractmethod\n    def get_game_status(self, game_id: str) -> GameStatus:\n        \"\"\"\n        Check if the game is over and get the winner if there's one.\n        \"\"\"\n        pass\n\n    @abstractmethod\n    def get_winner(self, game_id: str) -> str:\n        \"\"\"\n        Get the winner of the game.\n        \"\"\"\n        pass\n\n    @abstractmethod\n    def get_game(self) -> Game:\n        \"\"\"\n        Retrieve the state of the game.\n        \"\"\"\n        pass\n\n    @abstractmethod\n    def delete_game(self, game_id: str) -> None:\n        \"\"\"\n        Delete a game given its ID.\n        \"\"\"\n        pass\n\n    @abstractmethod\n    def create_game(self, game_id: str) -> None:\n        \"\"\"\n        Create a new game.\n        \"\"\"\n        pass\n\n```\nAt any moment you can run ```pytest``` to execute the tests.\nYou have two types of test: \n- positive tests => test the battleship game being used in ideal conditions\n- negative tests => tests the battleship game behaviour when used incorrectly\n\nSuccess criteria:\n- you will need to write a file called battleship.py that implements the abstract Battleship class.\n- this class will have to pass all the tests.\n- you're not allowed to modify any other file than the battleship.py. You can add other files as long as the main entrypoint is the battleship class.",
+      debugLog
+    );
+
+    expect(response.value.ok).toBe(true);
+    const sourceCode = agent.workspace.readFileSync("battleship.py");
     expect(sourceCode).toBeTruthy();
   });
 });
